Fixed tests for response status
