+++ tests/ResponseStatusTest.php
@@ -4,7 +4,6 @@ namespace Jasny\HttpMessage;
 
 use Jasny\HttpMessage\ResponseStatus;
 use PHPUnit_Framework_TestCase;
-use PHPUnit_Framework_MockObject_MockObject as MockObject;
 
 /**
  * @covers Jasny\HttpMessage\ResponseStatus
@@ -12,18 +11,13 @@ use PHPUnit_Framework_MockObject_MockObject as MockObject;
 class ResponseStatusTest extends PHPUnit_Framework_TestCase
 {
     /**
-     * @var ResponseStatus|MockObject
+     * @var ResponseStatus
      */
     protected $baseStatus;
     
     public function setUp()
     {
-        $this->baseStatus = $this->getMockBuilder(ResponseStatus::class)
-            ->setConstructorArgs(['1.1'])
-            ->setMethods(['header', 'headersSent', 'httpResponseCode'])
-            ->getMock();
-        
-        $this->baseStatus->method('headersSent')->willReturn(false);
+        $this->baseStatus = new ResponseStatus();
     }
     
     
@@ -52,7 +46,23 @@ class ResponseStatusTest extends PHPUnit_Framework_TestCase
      */
     public function testConstruct($status, $phrase, $expectPhrase)
     {
-        $responseStatus = new ResponseStatus('1.1', $status, $phrase);
+        $responseStatus = new ResponseStatus($status, $phrase);
+        
+        $this->assertSame($status, $responseStatus->getStatusCode());
+        $this->assertSame($expectPhrase, $responseStatus->getReasonPhrase());
+    }    
+
+    /**
+     * @dataProvider statusCodeProvider
+     * 
+     * @param int    $status
+     * @param string $phrase
+     * @param string $expectPhrase
+     */
+    public function testConstructWithResposnes($status, $phrase, $expectPhrase)
+    {
+        $original = new ResponseStatus($status, $phrase);
+        $responseStatus = new ResponseStatus($original);
         
         $this->assertSame($status, $responseStatus->getStatusCode());
         $this->assertSame($expectPhrase, $responseStatus->getReasonPhrase());
@@ -74,124 +84,6 @@ class ResponseStatusTest extends PHPUnit_Framework_TestCase
         
         $this->assertSame($status, $responseStatus->getStatusCode());
         $this->assertSame($expectPhrase, $responseStatus->getReasonPhrase());
-        
-        $this->assertFalse($this->baseStatus->isStale());
-    }
-    
-    public function testWithStatusSame()
-    {
-        $status = $this->baseStatus->withStatus(200);
-        $finalStatus = $status->withStatus(200);
-        
-        $this->assertSame($status, $finalStatus);
-    }
-    
-    /**
-     * @dataProvider statusCodeProvider
-     * 
-     * @param int    $status
-     * @param string $phrase
-     * @param string $expectPhrase
-     */
-    public function testWithStatusGlobal($status, $phrase, $expectPhrase)
-    {
-        $globalStatus = 200;
-        
-        $this->baseStatus->expects($this->any())->method('httpResponseCode')->willReturnReference($globalStatus);
-        $this->baseStatus->expects($this->once())->method('header')
-            ->with("HTTP/1.1 {$status} {$expectPhrase}")
-            ->will($this->returnCallback(function() use (&$globalStatus, $status) {
-                $globalStatus = $status;
-            }));
-        
-        $this->baseStatus->useGlobally();
-        
-        $responseStatus = $this->baseStatus->withStatus($status, $phrase);
-        
-        $this->assertInstanceOf(ResponseStatus::class, $responseStatus);
-        $this->assertNotSame($this->baseStatus, $responseStatus);
-
-        $this->assertSame($status, $responseStatus->getStatusCode());
-        $this->assertSame($expectPhrase, $responseStatus->getReasonPhrase());
-        
-        $this->assertTrue($this->baseStatus->isStale());
-    }
-    
-    public function testGetStatusGlobal()
-    {
-        $this->baseStatus->expects($this->any())->method('httpResponseCode')->willReturn(400);
-        $this->baseStatus->useGlobally();
-        
-        $this->assertSame(400, $this->baseStatus->getStatusCode());
-        $this->assertSame('Bad Request', $this->baseStatus->getReasonPhrase());
-    }
-    
-    
-    public function statusProvider()
-    {
-        return [
-            [200, 'OK'],
-            [400, 'Bad Request'],
-            [600, '']
-        ];
-    }
-
-    /**
-     * @dataProvider statusProvider
-     * 
-     * @param int    $status
-     * @param string $phrase
-     */
-    public function testUseLocally($status, $phrase)
-    {
-        $this->baseStatus->useGlobally();
-        
-        $this->baseStatus->expects($this->once())->method('httpResponseCode')->willReturn($status);
-        $this->baseStatus->useLocally();
-        
-        $this->assertEquals($status, $this->baseStatus->getStatusCode());
-        $this->assertSame($phrase, $this->baseStatus->getReasonPhrase());
-        
-        $this->baseStatus->useLocally();
-        
-        $this->assertFalse($this->baseStatus->isGlobal());
-    }
-    
-    public function testUseLocallyNoStatus()
-    {
-        $this->baseStatus->useGlobally();
-        $status = $this->baseStatus->withStatus(400);
-        
-        $status->expects($this->once())->method('httpResponseCode')->willReturn(false);
-        $status->useLocally();
-        
-        $this->assertEquals(200, $status->getStatusCode());
-        $this->assertSame('OK', $status->getReasonPhrase());
-        
-        $status->useLocally();
-        
-        $this->assertFalse($status->isGlobal());
-    }
-    
-    public function testUseGlobally()
-    {
-        $this->baseStatus->expects($this->once())->method('header')->with("HTTP/1.1 201 Created");
-        
-        $status = $this->baseStatus->withStatus(201);
-        $status->useGlobally();
-        
-        $this->assertTrue($status->isGlobal());
-    }
-    
-    public function testUseGloballyWhenStale()
-    {
-        $this->baseStatus->useGlobally();
-        $this->baseStatus->withStatus(200);
-        
-        $this->baseStatus->useGlobally();
-        $this->assertTrue($this->baseStatus->isStale());
-        
-        $this->assertTrue($this->baseStatus->isGlobal());
     }
     
     
@@ -233,7 +125,7 @@ class ResponseStatusTest extends PHPUnit_Framework_TestCase
      */
     public function testConstructInvalidTypeStatusCode()
     {
-        new ResponseStatus('1.1', 1020.20);
+        new ResponseStatus(1020.20);
     }
 
     /**
@@ -241,7 +133,7 @@ class ResponseStatusTest extends PHPUnit_Framework_TestCase
      */
     public function testConstructInvalidValueStatusCode()
     {
-        new ResponseStatus('1.1', 1020);
+        new ResponseStatus(1020);
     }
 
     /**
@@ -249,38 +141,6 @@ class ResponseStatusTest extends PHPUnit_Framework_TestCase
      */
     public function testConstructInvalidValueStatusPhrase()
     {
-        new ResponseStatus('1.1', 200, ['foo', 'bar']);
-    }
-    
-    
-    /**
-     * @expectedException \RuntimeException
-     * @expectedExceptionMessage Can not change stale object
-     */
-    public function testModifyStale()
-    {
-        $this->baseStatus->useGlobally();
-        
-        $this->baseStatus->withStatus(200);
-        $this->baseStatus->withStatus(400);
-    }
-    
-    
-    public function testWithProtocolVersion()
-    {
-        $this->baseStatus->expects($this->once())->method('header')->with("HTTP/2 200 OK");
-        
-        $responseStatus = $this->baseStatus->withProtocolVersion('2');
-        $responseStatus->useGlobally();
-        
-        $responseStatus->withStatus(200);
-    }
-    
-    /**
-     * @expectedException InvalidArgumentException
-     */
-    public function testWithProtocolVerionsInvalidArgument()
-    {
-        $this->baseStatus->withProtocolVersion(0);
+        new ResponseStatus(200, ['foo', 'bar']);
     }
 }
